requirements.txt
pygrok
regex
configparser
argparse
colorama
importlib
glob
console_progressbar
neo4j
virustotal-api
netaddr